TCP/UART robustness improvements
Now send the rx_len on the TCP socket as the first 4 bytes, so the
client knows how many bytes it can expect. This is necessary to stop
reads blocking forever on the client side.

Also increase the UART timeout to be more than a single scheduler tick.

Lastly, tune down the TCP transaction rx_len so that it's correct for a
LOGS command, hopefully reducing the amount of time we spend blocking.
This won't work for other commands, but for now it's an improvement.